refactor: standardize `rci` variable type across project files
- Rename `capability_index` to `rci` in `simulation.yaml`.
- Change `rci` type from `int8_t` to `int` in `robot.hpp`.
- Update `rci` type to `int` in `simulator.cpp`.

+++ include/farmbot_flatlands/robot.hpp
@@ -68,7 +68,7 @@ namespace robo {
         std::string ns;
         int identifier;
         std::string uuid;
-        int8_t rci;
+        int rci;
         Sensors sensors;
     };
